linked PhysicsCharacter

+++ include/Game/MathHelpers.h
@@ -18,7 +18,7 @@ static inline u16 nlVector3ToAngle(const nlVector3& dir, u16 offset = 0)
  */
 static inline u16 RadToAng16(f32 r)
 {
-    return (u16)(f32)(10430.378f * r);
+    return (u16)(s32)(10430.378f * r);
 }
 
 /*
